Einführung von ganzzahligen Größen

+++ include/zollstock/definition_helpers.hpp
@@ -8,7 +8,7 @@
 #define ZOLLSTOCK_DEFINE_LITERAL(symbol)                                                    \
     [[nodiscard]] consteval auto operator""_##symbol(unsigned long long int value) noexcept \
     {                                                                                       \
-        return quantity<symbol, double>{ value };                                           \
+        return quantity<symbol, int>{ value };                                              \
     }                                                                                       \
                                                                                             \
     [[nodiscard]] consteval auto operator""_##symbol(long double value) noexcept            \
